feat: #436 add ThemeSwitcher to storybook component
